requirements.txt
darkdetect~=0.8.0
pystray~=0.19.5
pillow~=10.4.0
python-rtmidi~=1.5.8
steelseries-sonar-py~=1.2.0
PySide6~=6.7.2